Add files via upload
In this program, I have implemented both the brute-force and recursive algorithms for the maximum subarray problem. Later on, I found size n which gives the crossover point at which the recursive algorithm beats the brute-force algorithm. Then, I changed the base case of the recursive algorithm to use the brute-force algorithm whenever the problem size is less than n0.